+++ src/Curiosity/Runtime.hs
@@ -42,9 +42,6 @@ module Curiosity.Runtime
   , writeCreateQuotationForm
   , submitCreateQuotationForm
   , setQuotationAsSignedFull
-  , filterQuotations
-  , filterQuotations'
-  , selectQuotationById
   -- ** Orders
   , filterOrders
   , filterOrders'
@@ -78,6 +75,8 @@ module Curiosity.Runtime
   , module RErr
   , module RIO
   , module AppM
+  , module Runtime.Q
+  , module Runtime.Ord
   ) where
 
 import qualified Commence.Multilogging         as ML
@@ -103,8 +102,13 @@ import qualified Curiosity.Data.SimpleContract as SimpleContract
 import qualified Curiosity.Data.User           as User
 import           Curiosity.Runtime.Error       as RErr
 import           Curiosity.Runtime.IO          as RIO
-import           Curiosity.Runtime.IO.AppM      ( AppM, runAppM, runAppMSafe )
+import           Curiosity.Runtime.IO.AppM      ( AppM
+                                                , runAppM
+                                                , runAppMSafe
+                                                )
 import qualified Curiosity.Runtime.IO.AppM     as AppM
+import           Curiosity.Runtime.Order       as Runtime.Ord
+import           Curiosity.Runtime.Quotation   as Runtime.Q
 import           Curiosity.Runtime.Type        as RType
 import           Curiosity.STM.Helpers          ( atomicallyM )
 import qualified Data.Aeson.Text               as Aeson
@@ -132,9 +136,8 @@ import           System.PosixCompat.Types       ( EpochTime )
 --------------------------------------------------------------------------------
 showThreads :: Threads -> IO [Text]
 showThreads ts = case ts of
-  NoThreads        -> pure ["Threads are disabled."]
-  ReplThreads mvarEmails ->
-    showThreads' "Email" mvarEmails
+  NoThreads                       -> pure ["Threads are disabled."]
+  ReplThreads mvarEmails          -> showThreads' "Email" mvarEmails
   HttpThreads mvarEmails mvarUnix -> do
     t1 <- showThreads' "Email" mvarEmails
     t2 <- showThreads' "UNIX-domain socket" mvarUnix
@@ -143,7 +146,9 @@ showThreads ts = case ts of
 showThreads' :: Text -> MVar ThreadId -> IO [Text]
 showThreads' name mvar = do
   b <- isEmptyMVar mvar
-  if b then pure [name <> " thread: stopped."] else pure [name <> " thread: running."]
+  if b
+    then pure [name <> " thread: stopped."]
+    else pure [name <> " thread: running."]
 
 
 --------------------------------------------------------------------------------
@@ -246,7 +251,7 @@ emptyReplThreads = do
 emptyHttpThreads :: IO Threads
 emptyHttpThreads = do
   mvarEmails <- newEmptyMVar
-  mvarUnix <- newEmptyMVar
+  mvarUnix   <- newEmptyMVar
   pure $ HttpThreads mvarEmails mvarUnix
 
 spawnEmailThread :: RunM Text
@@ -254,8 +259,8 @@ spawnEmailThread = do
   runtime <- ask
   ts      <- asks _rThreads
   case ts of
-    NoThreads        -> pure "Threads are disabled."
-    ReplThreads mvarEmails -> spawnEmailThread' runtime mvarEmails
+    NoThreads                -> pure "Threads are disabled."
+    ReplThreads mvarEmails   -> spawnEmailThread' runtime mvarEmails
     HttpThreads mvarEmails _ -> spawnEmailThread' runtime mvarEmails
 
 spawnEmailThread' :: Runtime -> MVar ThreadId -> RunM Text
@@ -277,8 +282,8 @@ killEmailThread :: RunM Text
 killEmailThread = do
   ts <- asks _rThreads
   case ts of
-    NoThreads        -> pure "Threads are disabled."
-    ReplThreads mvarEmails -> killEmailThread' mvarEmails
+    NoThreads                -> pure "Threads are disabled."
+    ReplThreads mvarEmails   -> killEmailThread' mvarEmails
     HttpThreads mvarEmails _ -> killEmailThread' mvarEmails
 
 killEmailThread' :: MVar ThreadId -> RunM Text
@@ -358,8 +363,8 @@ spawnUnixThread = do
   runtime <- ask
   ts      <- asks _rThreads
   case ts of
-    NoThreads        -> pure "Threads are disabled."
-    ReplThreads _ -> pure "No UNIX-domain socket thread in REPL." -- TODO
+    NoThreads              -> pure "Threads are disabled."
+    ReplThreads _          -> pure "No UNIX-domain socket thread in REPL." -- TODO
     HttpThreads _ mvarUnix -> spawnUnixThread' runtime mvarUnix
 
 spawnUnixThread' :: Runtime -> MVar ThreadId -> RunM Text
@@ -543,8 +548,7 @@ handleCommand runtime@Runtime {..} user command = do
             in  i <> " " <> n
       pure (ExitSuccess, map f profiles)
     Command.UpdateUser input -> do
-      output <- runAppMSafe runtime . atomicallyM $ Core.updateUser _rDb
-                                                                    input
+      output <- runAppMSafe runtime . atomicallyM $ Core.updateUser _rDb input
       case output of
         Right mid -> do
           case mid of
@@ -556,8 +560,7 @@ handleCommand runtime@Runtime {..} user command = do
             Left err -> pure (ExitFailure 1, [show err])
         Left err -> pure (ExitFailure 1, [show err])
     Command.DeleteUser input -> do
-      output <- runAppMSafe runtime . atomicallyM $ Core.deleteUser _rDb
-                                                                    input
+      output <- runAppMSafe runtime . atomicallyM $ Core.deleteUser _rDb input
       case output of
         Right mid -> do
           case mid of
@@ -940,64 +943,6 @@ linkBusinessUnitToUser'
 linkBusinessUnitToUser' slug uid role = do
   db <- asks _rDb
   atomicallyM $ Core.linkBusinessUnitToUser db slug uid role
-
-
---------------------------------------------------------------------------------
-createQuotation
-  :: Core.StmDb
-  -> Quotation.Quotation
-  -> STM (Either Quotation.Err Quotation.QuotationId)
-createQuotation db quotation = do
-  STM.catchSTM (Right <$> transaction) (pure . Left)
- where
-  transaction = do
-    newId <- Core.generateQuotationId db
-    let new = quotation { Quotation._quotationId = newId }
-    createQuotationFull db new >>= either STM.throwSTM pure
-
-createQuotationFull
-  :: Core.StmDb
-  -> Quotation.Quotation
-  -> STM (Either Quotation.Err Quotation.QuotationId)
-createQuotationFull db new = do
-  modifyQuotations db (++ [new])
-  pure . Right $ Quotation._quotationId new
-
-modifyQuotations
-  :: Core.StmDb -> ([Quotation.Quotation] -> [Quotation.Quotation]) -> STM ()
-modifyQuotations db f =
-  let tvar = Data._dbQuotations db in STM.modifyTVar tvar f
-
-createOrderForQuotation
-  :: Core.StmDb
-  -> (User.UserProfile, Quotation.QuotationId)
-     -- ^ TODO SignQuotation data type, including e.g. the signature data.
-  -> STM (Either Quotation.Err Order.OrderId)
-createOrderForQuotation db _ = do
-  mid <- createOrder db
-  case mid of
-    Right id              -> pure $ Right id
-    Left  (Order.Err err) -> pure $ Left $ Quotation.Err err
-
-
---------------------------------------------------------------------------------
-createOrder :: Core.StmDb -> STM (Either Order.Err Order.OrderId)
-createOrder db = do
-  STM.catchSTM (Right <$> transaction) (pure . Left)
- where
-  transaction = do
-    newId <- Core.generateOrderId db
-    let new = Order.Order newId
-    createOrderFull db new >>= either STM.throwSTM pure
-
-createOrderFull
-  :: Core.StmDb -> Order.Order -> STM (Either Order.Err Order.OrderId)
-createOrderFull db new = do
-  modifyOrders db (++ [new])
-  pure . Right $ Order._orderId new
-
-modifyOrders :: Core.StmDb -> ([Order.Order] -> [Order.Order]) -> STM ()
-modifyOrders db f = let tvar = Data._dbOrders db in STM.modifyTVar tvar f
 
 invoiceOrder
   :: Core.StmDb
@@ -1460,22 +1405,6 @@ setQuotationAsRejected db id mcomment = do
       _ -> pure . Left $ Quotation.Err "Quotation is not in the Sent state."
     Nothing -> pure . Left $ Quotation.Err "No such quotation."
 
-filterQuotations
-  :: Core.StmDb -> Quotation.Predicate -> STM [Quotation.Quotation]
-filterQuotations db predicate = do
-  let tvar = Data._dbQuotations db
-  records <- STM.readTVar tvar
-  pure $ filter (Quotation.applyPredicate predicate) records
-
-filterQuotations' :: Quotation.Predicate -> RunM [Quotation.Quotation]
-filterQuotations' predicate = do
-  db <- asks _rDb
-  atomicallyM $ filterQuotations db predicate
-
-selectQuotationById
-  :: Core.StmDb -> Quotation.QuotationId -> IO (Maybe Quotation.Quotation)
-selectQuotationById db id = STM.atomically $ Core.selectQuotationById db id
-
 
 --------------------------------------------------------------------------------
 filterOrders :: Core.StmDb -> Order.Predicate -> STM [Order.Order]
@@ -1719,8 +1648,7 @@ filterUsers' predicate = do
   db <- asks _rDb
   atomicallyM $ filterUsers db predicate
 
-signup
-  :: User.Signup -> RunM (Either User.Err (User.UserId, Email.EmailId))
+signup :: User.Signup -> RunM (Either User.Err (User.UserId, Email.EmailId))
 signup input = ML.localEnv (<> "Command" <> "Signup") $ do
   ML.info "Signing up new user..."
   db   <- asks _rDb
