+++ src/Curiosity/Runtime/Order.hs
@@ -1,0 +1,28 @@
+module Curiosity.Runtime.Order
+  ( createOrder
+  , createOrderFull
+  , modifyOrders
+  ) where
+
+import qualified Control.Concurrent.STM        as STM
+import qualified Curiosity.Core                as Core
+import qualified Curiosity.Data                as Data
+import qualified Curiosity.Data.Order          as Order
+
+createOrder :: Core.StmDb -> STM (Either Order.Err Order.OrderId)
+createOrder db = do
+  STM.catchSTM (Right <$> transaction) (pure . Left)
+ where
+  transaction = do
+    newId <- Core.generateOrderId db
+    let new = Order.Order newId
+    createOrderFull db new >>= either STM.throwSTM pure
+
+createOrderFull
+  :: Core.StmDb -> Order.Order -> STM (Either Order.Err Order.OrderId)
+createOrderFull db new =
+  modifyOrders db (<> [new]) $> Right (Order._orderId new)
+
+modifyOrders :: Core.StmDb -> ([Order.Order] -> [Order.Order]) -> STM ()
+modifyOrders db f = let tvar = Data._dbOrders db in STM.modifyTVar tvar f
+
